+++ src/app/background/index.js
@@ -7,14 +7,16 @@ import { initializeIntentCoherence } from '../../js/background/intentCoherence.j
 import { initializePomodoroRuntime } from '../../js/background/pomodoro.js';
 import { initializeReleaseBackupNoticeEligibility } from '../../js/background/releaseNotice.js';
 import { initializeScheduleMonitor } from '../../js/background/scheduleMonitor.js';
+import { addRuntimeMessageListener, getExtensionUrl } from '../../platform/chrome/runtime.js';
+import { createTab } from '../../platform/chrome/tabs.js';
 
 const contentBlockingRuntime = createContentBlockingBackgroundRuntime(chrome);
 
 chrome.action.onClicked.addListener(() => {
-  chrome.tabs.create({ url: chrome.runtime.getURL('src/options.html') });
+  createTab({ url: getExtensionUrl('src/options.html') }).catch(() => {});
 });
 
-chrome.runtime.onMessage.addListener(contentBlockingRuntime.handleRuntimeMessage);
+addRuntimeMessageListener(contentBlockingRuntime.handleRuntimeMessage);
 chrome.tabs.onUpdated.addListener(contentBlockingRuntime.handleTabUpdated);
 chrome.tabs.onRemoved.addListener(contentBlockingRuntime.handleTabRemoved);